tox-requirements-3.txt
# Base reqs
click>=6.7
databricks-cli>=0.14.3
# Test reqs
prospector[with_pyroma]==1.3.0
pylint==2.5.2
pep8-naming==0.5.0
pytest>=3.8.1
mock==2.0.0
decorator==4.2.1
rstcheck==3.2
pytest-cov==2.5.1
# Pin docutils version to address https://sourceforge.net/p/docutils/bugs/365/
docutils==0.14
codecov
requests_mock==1.5.2
pytest-mock==3.6.1